fix(app): centralize attachment pill content
