+++ packages/app/src/composer/draft/create-flow.test.ts
@@ -95,6 +95,7 @@ describe("useDraftAgentCreateFlow", () => {
       attachment: {
         type: "text",
         mimeType: "text/plain",
+        contextKind: "chat_history",
         title: "Chat history",
         text: "Previous conversation",
       },
